Simplify CI target matching and sidecar cleanup flow
